优化了很多代码。

+++ WIFICity/Manager/WifiManager/WIFIValidateInfo.h
@@ -11,6 +11,6 @@
 @interface WIFIValidateInfo : WIModel
 
 @property (nonatomic,copy)NSString *routIp;
-@property (nonatomic,assign)float expireTime;
+@property (nonatomic,assign)NSInteger expireTime;
 
 @end
